# [PM-6163] Speed up list items command - Only call GetGlobals() when we actually need them