@luckman212 2024-05-11
Is this landed yet? I tested it with a prerelease artifact from 2024.5.0 and am still seeing extremely long `list items` runtimes, see https://github.com/bitwarden/clients/issues/9142#issuecomment-2106007166