@oysandvik94 2024-05-24
Pretty slow for me too. I notice that i hangs a little bit after it prints out the results, so it's not getting the actual items that are slow. I feel like this is slow as well: `bw unlock --session $BW_SESSION --check`, around 800 ms. I run this in a script to check that my BW_SESSION is valid before I call `list items`, so together it's pretty slow.